Consolidate _no_real_repair and _agent_commands_available into conftest.py
These two autouse fixtures were duplicated across 15 test modules. Moving
them to tests/conftest.py lets pytest apply them automatically to all test
files without any per-module repetition.

Closes #443
