+++ tests/conftest.py
@@ -1,0 +1,34 @@
+from unittest.mock import patch
+
+import pytest
+
+
+@pytest.fixture(autouse=True)
+def _no_real_repair():
+    """Prevent attempt_repair from calling the real Gemini CLI in all tests.
+
+    Tests that explicitly test repair behaviour patch the orchestrator-level
+    import themselves, which takes precedence over this fixture.  Unit tests
+    for attempt_repair itself patch subprocess.run directly, so they are
+    unaffected here.
+    """
+    with patch("coding_review_agent_loop.orchestrator.attempt_repair", return_value=None):
+        yield
+
+
+@pytest.fixture(autouse=True)
+def _agent_commands_available(monkeypatch):
+    """Keep config tests independent of agent CLIs installed on the test host."""
+    import coding_review_agent_loop.config as config_module
+
+    real_which = config_module.shutil.which
+
+    def which(command):
+        resolved = real_which(command)
+        if resolved is not None:
+            return resolved
+        if command in {"claude", "codex", "gemini", "agy"}:
+            return f"/mock/bin/{command}"
+        return None
+
+    monkeypatch.setattr(config_module.shutil, "which", which)
