+++ tests/test_decomposition.py
@@ -1,5 +1,3 @@
-from unittest.mock import patch
-
 import pytest
 
 from coding_review_agent_loop.cli import AgentLoopError, run_issue_loop
@@ -17,35 +15,6 @@ from coding_review_agent_loop.github import IssueComment
 from coding_review_agent_loop.orchestrator import PostedRoundMetadata, _attach_round_metadata, _plan_subject
 from agent_loop_helpers import FakeRunner, make_config, plan_decomposition_json, structured_plan_review
 
-@pytest.fixture(autouse=True)
-def _no_real_repair():
-    """Prevent attempt_repair from calling the real Gemini CLI in all tests.
-
-    Tests that explicitly test repair behaviour patch the orchestrator-level
-    import themselves, which takes precedence over this fixture.  Unit tests
-    for attempt_repair itself patch subprocess.run directly, so they are
-    unaffected here.
-    """
-    with patch("coding_review_agent_loop.orchestrator.attempt_repair", return_value=None):
-        yield
-
-
-@pytest.fixture(autouse=True)
-def _agent_commands_available(monkeypatch):
-    """Keep config tests independent of agent CLIs installed on the test host."""
-    import coding_review_agent_loop.config as config_module
-
-    real_which = config_module.shutil.which
-
-    def which(command):
-        resolved = real_which(command)
-        if resolved is not None:
-            return resolved
-        if command in {"claude", "codex", "gemini", "agy"}:
-            return f"/mock/bin/{command}"
-        return None
-
-    monkeypatch.setattr(config_module.shutil, "which", which)
 
 def test_parse_plan_decomposition_accepts_agent_and_human_phases():
     parsed = parse_plan_decomposition(
